Synch with H2.0.3 - added Identifer for UUIDHexGenerator and GuidGenerator.  Fixed UUIDStringGenerator to work like doc says it will.  Other changes were code formatting related.
SVN: trunk@324

+++ src/NHibernate/Id/CounterGenerator.cs
@@ -1,25 +1,31 @@
 using System;
 using NHibernate.Engine;
 
-namespace NHibernate.Id {
+namespace NHibernate.Id 
+{
 	/// <summary>
-	/// An <c>IIdentifierGenerator</c> that returns a <c>long</c> constructed from the system
+	/// An <c>IIdentifierGenerator</c> that returns a <c>Int64</c> constructed from the system
 	/// time and a counter value. Not safe for use in a clustser!
 	/// </summary>
-	public class CounterGenerator : IIdentifierGenerator {
+	public class CounterGenerator : IIdentifierGenerator 
+	{
 		
-		private static short counter = 0;
+		private static short counter = (short)0;
 
-		protected short Count {
-			get {
-				lock(typeof(CounterGenerator)) {
+		protected short Count 
+		{
+			get 
+			{
+				lock(typeof(CounterGenerator)) 
+				{
 					if (counter<0) counter=0;
 					return counter++;
 				}
 			}
 		}
 
-		public object Generate(ISessionImplementor cache, object obj) {
+		public object Generate(ISessionImplementor cache, object obj) 
+		{
 			return ( DateTime.Now.Ticks << 16 ) + Count;
 		}
